+++ models.py
@@ -259,6 +259,7 @@ class AuthEventType(models.TextChoices):
     PASSKEY_REGISTERED  = 'passkey_registered'
     PASSKEY_LOGIN       = 'passkey_login'
     PASSKEY_REMOVED     = 'passkey_removed'
+    PASSKEY_RENAMED     = 'passkey_renamed'
     OAUTH_LOGIN         = 'oauth_login'
     QR_LOGIN            = 'qr_login'
     SSO_LOGIN           = 'sso_login'